Entire rewrite of the project (simple-validation)
I have rewrittent the class from scratch. It now validates all forms on an
include-all rule and in order to not validate a form, you can either add
the class [novalidate] or the attribute [novalidate] to the form element.
Likewise, you can do the same for any element inside a form. Add the class
[novalidate] or the attribute [novalidate] and that input will not be
validated with the rest of the form elements. I plan to add more support
for other elements as well, but it currently only supports text and email
inputs as that is what was supported by Chad Humphrey prior to my
engagement. I hope you find this more useful!