Adding files for a github page.
These files will be the resources necessary to run a github page for this
repo. After creating the gh-pages branch, we can have a page set up to
automatically build when we push changes to it (it is another branch
though). This way, we can have a live site to demonstrate this library to
everyone and link to it from within the repo and the readme.