Another major refactoring of the code.
Now it is set up a lot more like a library, and it even allows you re
configure some settings now. The entire SimpleValidator namespace is now
set to window so it can be accessed through window.SimpleValidator. This
also allows you to configure it anywhere from changing the regular
expression used to changes the form selectors. This will make it much more
manageable and also it just looks a lot cleaner.