A little more code refactoring...
I know it seems like overkill, but now the code is factored out into a
namespace pretty well. This will allow for a huge opportunity of
customization in the future without packing extra features in this
library. This will keep it light but also allow for devs to pack in and
use more features such as the validation helper methods. There are two
main calls to make now: SimpleValidator.Initialize() which sets all event
handlers including the form submit event handler, and also
SimpleValidator.Validate() which forces a global validation at the moment.
That means all forms and all formElements that aren't set to novalidate
will validate immediately. If you want to validate just one form randomly
or even just validate a subset of forms, call the helper method
SimpleValidator.Helpers.ValidateForm() or
SimpleValidator.Helpers.ValidateForms().