Added support for required attribute.
If an input has the [required] attribute, simple-validation.js will now
make sure that the input is not just an empty string. If it is, it will
mark it as invalid. If an input is not required, however, the validator
will allow it as an empty string (no input), but will still validate
against the other rules if the input is not empty. This allows us to
maintain validation even on top of optional inputs.

Takeaway: If you want an input to be required now, please mark it with a
[required] attribute in the html. I'm just trying to conform to html5
standards.