Added checkbox validation! Woohoo!
That was one of the validations I thought might be a bit more challenging.
It is finished though, and that means radio validation shouldn't be much
longer either! Currently, the way checkbox validation works is that every
checkbox has a validation event. If a checkbox has a name attribute, it
then has a family including all checkboxes under that form with the same
name. The validation then checks to see if any of the family has a
required attribute, and if so, the entire family becomes required to have
at least one box checked. If not, none of the boxes have to be checked.
Now, if the checkboxes are not a part of the same family, then checking
one will still not clear another if it is required as well. Also, if they
are not a part of the same family, requiring one will not require the
others. You may add the required attribute to all of the checkboxes under
the same family if you wish, but it isn't necessary for the validation to
work properly.